cloud_functions/market_regime_production/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.14.1
yfinance==0.2.33
pandas>=2.0.0
